# Loads IANA registry and use it for validation and canonicalization