Updates `LanguageTag::validate` and `LanguageTag::canonicalize` to use it.

`LanguageTag::canonicalize` returns a validation error if there is no simple way of returning a "canonical version" of the language tag. For example, it does not fail if the region is not in the registry but fails if there are multiple language extensions because there is no proper way to choose which one to use to update the primary language.

To update the IANA registry local version run `python3 scripts/update_from_iana_subtag_registery.py && cargo fmt`