@Tpt 2019-03-14
Benchmarks from PR #15:
master:
test canonicalize_tags      ... bench:      15,937 ns/iter (+/- 3,803) = 39 MB/s
test parse_well_formed_tags ... bench:       9,411 ns/iter (+/- 1,300) = 62 MB/s
test validate_tags          ... bench:       1,501 ns/iter (+/- 124) = 415 MB/s

this PR:
test canonicalize_tags      ... bench:      33,916 ns/iter (+/- 1,080) = 18 MB/s
test parse_well_formed_tags ... bench:       8,664 ns/iter (+/- 279) = 67 MB/s
test validate_tags          ... bench:      20,921 ns/iter (+/- 865) = 29 MB/s

This is a significant slowdown but many more features are now implemented