@pyfisch 2019-03-16
The registry file is *huge*. How big is the resulting executable?