@Tpt 2019-03-16
The *.rlib file is 8MB but release executables are much smaller (~4MB for the benchmarks)