@pyfisch 2019-03-16
The release tests after running strip have a size of 1.5M (before 698 kB). I've decided for myself that this is too much and experimented a bit with more efficient storage. I will publish a PR for you to review soon. With this optimization the executable should be around 900 kB and the benchmarks will be a bit worse.