@pyfisch 2019-03-17
Hi,

I've coded myself a bit into a corner. While I got the executable size down the performance is also reduced quite a bit. The main problem is that each needs a pointer and a length (16 bytes) and therefore a `LanguageRecord` has a size of 64 bytes on 64 bit machines (the options are free). Additionally the string data needs to be stored. On the other hand the binary lookup by tag should be fast.

```rust
pub struct LanguageRecord {
    pub subtag: &'static str,
    pub preferred_value: Option<&'static str>,
    pub suppress_script: Option<&'static str>,
    pub macrolanguage: Option<&'static str>
}
```