@Tpt 2019-03-17
Thanks!

There is probably a way to improve the storage space by avoiding `&'static str`. For most of the record types, the subtag size is fixed so we could use a [u8; MAX_SIZE] as container. It may require some unsafe code to avoid a lot of UTF-8 validity checking but it could be much smaller and hopefully faster because of less indirections.

For example, `LanguageRecord` would be:

```rust
pub struct LanguageRecord {
    pub subtag: [u8; 3], // The last byte could be \0 for tags of length 2 like "en"
    pub preferred_value: Option<[u8; 3]>,
    pub suppress_script: Option<[u8; 4]>,
    pub macrolanguage: Option<[u8; 2]>
}

```

We may also experiment `repr(packed)` on it.