@pyfisch 2019-03-17
Sound good. Yeah it probably a good idea to store the fixed fields in arrays. Is there a way to get a 'static str from an array? (How does the borrowing work?)

Do you want to write a PR that improves storage space?