@Tpt 2019-03-17
> Is there a way to get a 'static str from an array?

Yes, the good function is [str::from_utf8()](https://doc.rust-lang.org/std/str/fn.from_utf8.html) that takes a `&'a [u8]` as input and give back a `&'a str`. The last example of the doc is exactly what we are looking for I think.

I could try to work on it but if you whish to do so, feel free to do it.