@pyfisch 2019-03-17
Ah I see why I can't use `'static str` with LanguageRecord but the `&'a str` you suggest is good enough.

> I could try to work on it but if you whish to do so, feel free to do it.

I would be great if you work on it.